# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface GeminiConversation {
  id: number;
  title: string;
  role: string;
  createdAt: string;
}

export interface GeminiMessage {
  id: number;
  conversationId: number;
  role: string;
  content: string;
  createdAt: string;
}

export interface GeminiConversationInput {
  title: string;
  role: string;
}

export interface GeminiMessageInput {
  content: string;
}

export interface GeminiConversationWithMessages {
  id: number;
  title: string;
  role: string;
  createdAt: string;
  messages: GeminiMessage[];
}

export interface GeminiError {
  error: string;
}